+++ tests/server/disabled.c
@@ -116,9 +116,6 @@ static const char *disabled[]={
   "win32-ca-search-safe",
 #endif
 #endif
-#ifndef USE_HTTPSRR
-  "HTTPSRR",
-#endif
   NULL
 };
